+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/plugin/state/querychannels/ChannelsStateData.kt
@@ -22,7 +22,9 @@ public sealed class ChannelsStateData {
     /** No query is currently running.
      * If you know that a query will be started you typically want to display a loading icon.
      */
-    public object NoQueryActive : ChannelsStateData()
+    public object NoQueryActive : ChannelsStateData() {
+        override fun toString(): String = "ChannelsStateData.NoQueryActive"
+    }
 
     /** Indicates we are loading the first page of results.
      * We are in this state if QueryChannelsState.loading is true
@@ -31,13 +33,21 @@ public sealed class ChannelsStateData {
      * @see QueryChannelsState.loadingMore
      * @see QueryChannelsState.loading
      */
-    public object Loading : ChannelsStateData()
+    public object Loading : ChannelsStateData() {
+        override fun toString(): String = "ChannelsStateData.Loading"
+    }
 
     /** If we are offline and don't have channels stored in offline storage, typically displayed as an error condition. */
-    public object OfflineNoResults : ChannelsStateData()
+    public object OfflineNoResults : ChannelsStateData() {
+        override fun toString(): String = "ChannelsStateData.OfflineNoResults"
+    }
 
     /** The list of channels, loaded either from offline storage or an API call.
      * Observe chatDomain.online to know if results are currently up to date
      */
-    public data class Result(val channels: List<Channel>) : ChannelsStateData()
+    public data class Result(val channels: List<Channel>) : ChannelsStateData() {
+        override fun toString(): String {
+            return "ChannelsStateData.Result(channels.size=${channels.size})"
+        }
+    }
 }