+++ CHANGELOG.md
@@ -60,6 +60,7 @@
 ## stream-chat-android-ui-components
 ### 🐞 Fixed
 - Fixed the display of disconnected state in channel list and message list headers. [#3943](https://github.com/GetStream/stream-chat-android/pull/3943)
+- Fixed list state race condition while switching filters in channel list. [#3939](https://github.com/GetStream/stream-chat-android/pull/3939/files)
 
 ### ⬆️ Improved